tests/wpt/tests/html/webappapis/dynamic-markup-insertion/document-write/script_006.html

<!doctype html>
<title>document.write external script followed by internal script</title>
<script src="/resources/testharness.js"></script><script src="/resources/testharnessreport.js"></script>
<script>
var t = async_test();
var order = [];
t.step(function() {
  order.push(1);
  document.write("<script src='006.js'><"+"/script><script>t.step(function(){order.push(4)})</script"+">");
  order.push(2);
});
</script>
<script>
t.step(function() {
  order.push(5);
  assert_array_equals(order, [1,2,3,4,5]);
});
t.done();
</script>
<div id="log"></div>
